VERSION FINAL, falta documentacion interna

+++ Herramienta de Alto Nivel/src/GUI/FrameTimer.java
@@ -244,49 +244,64 @@ public class FrameTimer extends javax.swing.JFrame {
             Integer.parseInt(campoBase.getText());
             FileWriter fichero = null;
             PrintWriter pw = null;
-            try
-            {
-                String numTimer = String.valueOf(main.cuentaTimer);
-                fichero = new FileWriter("grafo.XML", true);
-                pw = new PrintWriter(fichero);
-                pw.println("<module type='timer_" + numTimer + "'>");
-                pw.println("    <datacounter>" + campoConteo.getSelectedItem().toString() + "</datacounter>");
-                pw.println("    <dataperiod>" + campoPeriodo.getText() + "</dataperiod>");
-                pw.println("    <unit>" + campoUnit.getSelectedItem().toString() + "</unit>");
-                if (campoEditable.isSelected())
-                    pw.println("    <editable>" + "1" + "</editable>");
-                else
-                    pw.println("    <editable>" + "0" + "</editable>");
-                pw.println("    <direction>" + campoBase.getText() + "</direction>");
-                pw.println("</module>");
-                int base = Integer.parseInt(campoBase.getText());
-                main.agregaDireccion("" + base + "_" + (base + 1));
-                main.direccionTimer = campoBase.getText();
-                main.pintaTimer();
+            if(existeDireccion()){
+                JOptionPane.showMessageDialog(null,"La base no es valida");
+            }else{
+                try
+                {
+                    String numTimer = String.valueOf(main.cuentaTimer);
+                    fichero = new FileWriter("grafo.XML", true);
+                    pw = new PrintWriter(fichero);
+                    pw.println("<module type='timer_" + numTimer + "'>");
+                    pw.println("    <counter>" + campoConteo.getSelectedItem().toString() + "</counter>");
+                    pw.println("    <period>" + campoPeriodo.getText() + "</period>");
+                    pw.println("    <unit>" + campoUnit.getSelectedItem().toString() + "</unit>");
+                    if (campoEditable.isSelected())
+                        pw.println("    <editable>" + "1" + "</editable>");
+                    else
+                        pw.println("    <editable>" + "0" + "</editable>");
+                    pw.println("    <direction>" + campoBase.getText() + "</direction>");
+                    pw.println("</module>");
+                    int base = Integer.parseInt(campoBase.getText());
+                    main.agregaDireccion("" + base + "_" + (base + 1));
+                    main.direccionTimer = campoBase.getText();
+                    main.pintaTimer();
 
-            } catch (Exception e) {
-                e.printStackTrace();
-            } finally {
-               try {
-               // Nuevamente aprovechamos el finally para 
-               // asegurarnos que se cierra el fichero.
-               if (null != fichero)
-                  fichero.close();
-               } catch (Exception e2) {
-                  e2.printStackTrace();
-               }
-            }
-            dispose();
-	} 
+                } catch (Exception e) {
+                    e.printStackTrace();
+                } finally {
+                   try {
+                   // Nuevamente aprovechamos el finally para 
+                   // asegurarnos que se cierra el fichero.
+                   if (null != fichero)
+                      fichero.close();
+                   } catch (Exception e2) {
+                      e2.printStackTrace();
+                   }
+                }
+                dispose();
+            } 
+        }
         catch (NumberFormatException nfe){
             JOptionPane.showMessageDialog(null,"Debe ingresar valores numericos"); 
 	}
         
-        
-        
-        
     }//GEN-LAST:event_jButton1ActionPerformed
 
+    public boolean existeDireccion(){
+        int i = 0;
+        while(main.dirBase[i] != null){
+            if (main.dirBase[i].equals(campoBase.getText())){
+                return true;
+            }
+            i++;
+        }
+        
+        return false;
+    }
+    
+    
+    
     private void jButton2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton2ActionPerformed
         // TODO add your handling code here:
         dispose();
